+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/domain/GitHubNotificationRepository.kt
@@ -2,6 +2,6 @@ package com.github.naoyukik.intellijplugingithubnotifications.domain
 
 import com.github.naoyukik.intellijplugingithubnotifications.domain.model.GitHubNotification
 
-interface NotificationRepository {
+interface GitHubNotificationRepository {
     fun fetchNotifications(): List<GitHubNotification>
 }