+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/userInterface/GitHubNotificationsToolWindowFactory.kt
@@ -1,9 +1,10 @@
 package com.github.naoyukik.intellijplugingithubnotifications.userInterface
 
 import com.github.naoyukik.intellijplugingithubnotifications.applicaton.ApiClientWorkflow
+import com.github.naoyukik.intellijplugingithubnotifications.applicaton.NotificationWorkflow
 import com.github.naoyukik.intellijplugingithubnotifications.applicaton.SettingStateWorkflow
 import com.github.naoyukik.intellijplugingithubnotifications.applicaton.dto.TableDataDto
-import com.github.naoyukik.intellijplugingithubnotifications.infrastructure.NotificationRepositoryImpl
+import com.github.naoyukik.intellijplugingithubnotifications.infrastructure.GitHubNotificationRepositoryImpl
 import com.github.naoyukik.intellijplugingithubnotifications.infrastructure.SettingStateRepositoryImpl
 import com.github.naoyukik.intellijplugingithubnotifications.utility.DateTimeHandler
 import com.intellij.icons.AllIcons
@@ -33,7 +34,6 @@ import java.awt.event.MouseEvent
 import java.io.IOException
 import java.net.URI
 import java.net.URISyntaxException
-import java.time.LocalDateTime
 import java.util.Timer
 import javax.swing.JComponent
 import javax.swing.table.DefaultTableModel
@@ -43,7 +43,7 @@ import kotlin.coroutines.CoroutineContext
 
 @Suppress("TooManyFunctions")
 class GitHubNotificationsToolWindowFactory : ToolWindowFactory, DumbAware, CoroutineScope, Disposable {
-    private val apiClientWorkflow = ApiClientWorkflow(NotificationRepositoryImpl())
+    private val apiClientWorkflow = ApiClientWorkflow(GitHubNotificationRepositoryImpl())
     private val settingStateWorkflow = SettingStateWorkflow(SettingStateRepositoryImpl())
     private val coroutineJob = Job()
     private var timer: Timer? = null
@@ -74,7 +74,7 @@ class GitHubNotificationsToolWindowFactory : ToolWindowFactory, DumbAware, Corou
 
         val settingState = settingStateWorkflow.getFetchInterval()
 
-        startAutoRefresh(notificationToolTable, settingState.fetchInterval)
+        startAutoRefresh(notificationToolTable, settingState.fetchInterval, project)
     }
 
     private fun createActionToolbar(actionGroup: DefaultActionGroup, targetComponent: JComponent): ActionToolbar {
@@ -173,14 +173,14 @@ class GitHubNotificationsToolWindowFactory : ToolWindowFactory, DumbAware, Corou
         })
     }
 
-    private fun startAutoRefresh(table: JBTable, minute: Int) {
+    private fun startAutoRefresh(table: JBTable, minute: Int, project: Project) {
         timer = fixedRateTimer(
             name = "GitHubNotificationRefresher",
             initialDelay = 0L,
             period = (minute * 60 * 1000).toLong(),
         ) {
             refreshNotifications(table)
-            println(LocalDateTime.now())
+            NotificationWorkflow().fetchedNotification(project)
         }
     }